feat(critical): Implement critical safety fixes and important improvements
- CRITICAL-001: Implement circuit breaker check in executor
- CRITICAL-002: Add atomic check-and-reserve for idempotency
- CRITICAL-003: Harden position sizer against NaN/Inf
- CRITICAL-004: Fix async/await inconsistencies in adapters
- IMPORTANT-001: Add SignalAdapterService
- IMPORTANT-002: Add concurrent stress tests
- IMPORTANT-003: Enforce strict feature schema validation
- IMPORTANT-004: Add ModelOutputMonitor
- IMPORTANT-005: Add configuration validation (stake limits)

+++ tests/data/test_validation.py
@@ -28,9 +28,9 @@ class TestFeatureBuilderValidation(unittest.TestCase):
             [103.0, 104.0, 102.0, 102.5, 1200.0, 10120.0],
             [102.5, 106.0, 102.0, 105.0, 1800.0, 10180.0],
             [105.0, 107.0, 104.0, 106.0, 2000.0, 10240.0]
-        ])
+        ], dtype=np.float32)
         
-        self.valid_ticks = np.ones(10) * 100.0
+        self.valid_ticks = (np.ones(10) * 100.0).astype(np.float32)
 
     def test_valid_data(self):
         """Test valid data passes."""
